+++ docs/ReleaseNotes.md
@@ -12,6 +12,7 @@ New Features in 3.70.0pre
 
 Bugs Fixed in 3.70.0pre
 -----------------------
+* Improved support for high contrast mode in Windows.
 * "Always on top" now persists correctly across restarts when "Start Password Safe at Login" is selected.
 * [GH1534](https://github.com/pwsafe/pwsafe/issues/1534) No longer offer unsupported traditional Chinese in NSIS installer.
 * [GH1531](https://github.com/pwsafe/pwsafe/issues/1531) Followup on [GH1524](https://github.com/pwsafe/pwsafe/issues/1524) for Toolbar action.